Drupal Instance
This tutorial shows you how to launch and configure a Drupal instance on Lightsail. It includes steps to connect to your instance by using SSH, signing in to your Drupal website, creating a static IP, attaching it to your instance, creating a DNS zone, and map your domain to your instance.